Updated the YMF262 code to work with int almost exclusively: - Changed RateBase and RateDivisor from uint to int. Changed _clock and _rate from uint to int. - Changed the type of OPLSlotOutArray from short to int. - Changed the phase generator Tick and Step fields to int. - Changed the output slots and phase modulation input latches from MutableBox<short> to MutableBox<int>. - Changed the wave table offset field from uint to int. - Changed the OPL3Channel FrequencyStepBase from uint to int. - Changed the type of FNumberTableArray from uint to int. - Changed uint fields in EnvelopeGeneratorRate from uint to int. - Changed TotalLevel and SustainLevel in EnvelopeGeneratorFields from uint to int. - Changed LFO state, Tick and Step in LowFrequencyOscillatorModulatorFields and CurrentAmplitude in LowFrequencyOscillatorFields from uint to int. Changed LFOShift to 23, since the values now butt up against the sign bit. - Changed BlockAndFunctionNumber and KeyScaleLevelBase in OPL3Channel from uint to int. - Changed the type of OutputControlValuesArray from uint to int. - Changed GlobalEnvelopeGeneratorState fields from uint to int. - Changed Value, Phase and Period in NoiseFields from uint to int. - In YMF262Tables.cs, changed all uint array types to int. - Simplified all associated math, removing casts and unchecked expressions. In some cases, an entire method was instead moved into an unchecked block. Reworked array objects in YMF262Chip.cs to use InlineArray structs instead: _channels, _outputMasks, _outputControlValues, _channelOutputSlots, _fnumberTable. Moved initialization of _channelOutputSlots to the constructor. Reworked array objects in YMF262Tables.cs to use InlineArray structs. Moved initialization to methods InitializeSineTable, InitializeTotalLevelTable, InitializeSlotByRegister, InitializeKeyScalelevelTable, InitializeKeyScaleLevelShift, InitializeSustainLevels, InitializeEnvelopeGeneratorIncrements, InitializeEnvelopeGeneratorRateSelect, InitializeEnvelopeGeneratorRateShift, InitializeMultiple, InitializeLFOAmplitudeModulatorTable and InitializeLFOPhaseModulatorTable. Fixed the OPL3Slot structure padding to correctly pad to a length of 128.

+++ QBX/Hardware/AdLib/OPLSlotOutArray.cs
@@ -5,5 +5,5 @@ namespace QBX.Hardware.AdLib;
 [InlineArray(2)]
 struct OPLSlotOutArray
 {
-	short _element0;
+	int _element0;
 }
